requirements.txt
--find-links https://pypi.python.org/simple/pywebview/

Django >= 1.8.4
pywebview >= 1.0 ; sys_platform == 'win32' and python_version >= '3.0'
pywebview >= 0.9 ; sys_platform != 'win32' or python_version < '3.0'

nose >= 1.3.7
